Moved optimized pool into jint namespace.
