+++ Jint/Runtime/DenyJsAccess.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 
-namespace CreateAR.EnkluPlayer.Scripting
+namespace Jint.Runtime
 {
     /// <summary>
     /// Restricts method from being called by Js.
     /// </summary>
     [AttributeUsage(AttributeTargets.Method | AttributeTargets.Property)]
-    public class DenyJsAccess : Attribute
-    {
-        //
-    }
+    public class DenyJsAccess : Attribute { }
 }
